move analyzePath to router
